Update renters.md
As discussed with the Dutch branch (Youran) it confused me that the policy describes 2 ways of business. Primarily business and whole business. Because more and more people in the primary target group (at least for the Netherlands) are using their smartphone and laptop privately and for business I asked for a statement. Youran told me it's about the invoice. When the invoice is paid by the business then it isn't covered. When paid privately it is covered. I hope my changes makes this more clear for everyone.